adventure_proj/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
djangorestframework==3.15.1
filelock==3.13.3
fsspec==2024.3.1
generativeai==0.0.1
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
huggingface-hub==0.22.2
idna==3.6
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
